+++ app/src/features/surveys/view/summary-results/NoSummaryResults.tsx
