+++ app/src/features/surveys/view/summary-results/SummaryResultsLoading.tsx
